Move 37 harness tests out of the e2e project into the clio.tests unit lane
More than half of the NoEnvironment tier finishes under 50 ms, and 31 of those
tests never mention an MCP server at all: they exercise the clio.mcp.e2e
harness itself (an appcmd XML parser, secret redaction, an envelope reader,
cleanup ownership, progress collection, worker-release waiting). They were
gated by the ~47-minute stand-backed TeamCity build, which is the only lane
that runs clio.mcp.e2e, so a broken string parser was reported 45 minutes late
and reported nothing at all whenever the stand was down.

Six fixtures (37 tests) now live in clio.tests/McpE2EHarness/ with
Category("Unit") and run in about a second. The audit treated this as blocked
on a choice between extracting a shared support assembly and leaving it alone;
a third option is cheaper than both, because clio.tests already references
clio.mcp.e2e for McpFixturePolicyTests: one InternalsVisibleTo("clio.tests")
keeps every Support type internal and needs no restructuring. Allure
attributes were dropped (clio.tests does not reference Allure.NUnit) and the
"E2E" suffix removed from the resolver fixture's name. No assertion changed.

McpWorkerWorkingDirectoryE2ETests stays, although it is just as fast: it
spawns a real child process and asserts the working directory that process
reports about itself. Speed is not the criterion; what the test starts is.

Also fixes a zero-skip regression that arrived from master: PR #1399 added a
Sandbox test to ListEntityClientSchemasToolE2ETests while the fixture kept a
class-level NoEnvironment category. NUnit categories are additive, so the fast
tier selected that test and skipped it for want of a stand, against the tier's
Skipped == 0 criterion. Per-method categories now, as in Stage 3.

Validated: dotnet test clio.tests -f net10.0 --filter "FullyQualifiedName~McpFixturePolicyTests|FullyQualifiedName~Clio.Tests.McpE2EHarness" (51/51, 1 s); full NoEnvironment tier 449 tests / 9m14s with the same container-specific failures as before and, after the category fix, 0 skips in the affected fixture; clio.mcp.e2e builds clean for net8.0 and net10.0
