+++ clio.tests/McpE2EHarness/ClioCliCommandRunnerRedactionTests.cs
@@ -1,7 +1,9 @@
+using System;
 using Clio.Mcp.E2E.Support.Configuration;
 using FluentAssertions;
+using NUnit.Framework;
 
-namespace Clio.Mcp.E2E;
+namespace Clio.Tests.McpE2EHarness;
 
 /// <summary>
 /// Env-free unit coverage for <see cref="ClioCliCommandRunner.RedactSecrets"/>: the diagnostics blocks
@@ -9,7 +11,7 @@ namespace Clio.Mcp.E2E;
 /// verbatim, so the value following a secret flag must never reach the (TeamCity-retained) CI log.
 /// </summary>
 [TestFixture]
-[Category("McpE2E.NoEnvironment")]
+[Category("Unit")]
 public sealed class ClioCliCommandRunnerRedactionTests {
 
 	[Test]
